webapp_chatterbox/requirements.txt
chatterbox-tts
flask
flask-cors
requests
torch
torchaudio
